requirements/base.txt
Django==1.8.14
Unipath==1.1
djangorestframework==3.4.7
django-money==0.9.1
py-moneyed==0.6.0
